Fixed primary issues
I'd made a mistake in manually calculating the cheapest holiday from Manchester
Re-wrote the query to correctly filter hotels by local airport - the previous join wasn't working right.
Removed a criteria field that was only serving to add confusion and complexity to the query and didn't provide additional benefits.
If the arrival date and departure date are ever different, we should add this back in.